Fixed the erroneous output in task 0.2

+++ src/com/umuzi/TaskTwo.java
@@ -3,13 +3,13 @@ package com.umuzi;
 public class TaskTwo {
 
     public void taskTwo() {
-        int x,y,z,a,b;
+        double x,y,z,a,b;
 
-        x = 1 + 1 * 2;
-        y = (1 + 1) * 2;
-        z = 1 + (1 * 2);
-        a = 1 + 1 * 2 / 2;
-        b = (1 + 1 * 2) / 2;
+        x = (double) 1 + 1 * 2;
+        y = (double) (1 + 1) * 2;
+        z = (double) 1 + (1 * 2);
+        a = (double) 1 + 1 * 2 / 2;
+        b = (double) (1 + 1 * 2) / 2;
 
         System.out.println(x + "\n" + y + "\n" + z + "\n" + a + "\n" + b);
     }
